refactor(moonrun): isolate async signal lifecycle
