+++ crates/moonrun/src/async_api/runtime.rs
@@ -18,7 +18,6 @@
 
 use super::context::ImportContext;
 use super::provenance::ported_imports;
-use crate::async_sys::signal;
 use crate::run_termination::RunTermination;
 
 pub(super) fn exit(context: &mut ImportContext<'_, '_>, code: i32) {
@@ -34,6 +33,9 @@ pub(super) fn terminate_process_by_signal(
     context: &mut ImportContext<'_, '_>,
     signal: i32,
 ) {
-    context.request_termination(signal::terminate_process_by_signal(signal))
+    // The native implementation terminates its process here. Moonrun instead
+    // records the equivalent outcome so the outer adapter can apply it after
+    // the Run has torn down.
+    context.request_termination(RunTermination::KilledBySignal(signal))
 }
 }
